Refactor firebase config
