+++ server/data-store/src/lib/auth/index.ts
@@ -1,28 +1,16 @@
-import * as firebase from "firebase/app";
-import "firebase/auth";
+import * as firebase from 'firebase/app';
+import 'firebase/auth';
+import firebaseConfig from '../constants'
 
+firebase.initializeApp(firebaseConfig);
 export interface Credential {
   email: string,
   password: string
 }
-
 interface firebaseError {
   code: string,
   message: string
 }
-
-const firebaseConfig = {
-  apiKey: "AIzaSyAOVXtaV1OsZp3YWwH_d3fiL2pdbGUuQH8",
-  authDomain: "growbud-50ed4.firebaseapp.com",
-  databaseURL: "https://growbud-50ed4.firebaseio.com",
-  projectId: "growbud-50ed4",
-  storageBucket: "growbud-50ed4.appspot.com",
-  messagingSenderId: "1032847847071",
-  appId: "1:1032847847071:web:116560c04ed237deb9bcaa",
-  measurementId: "G-N5LWZZRVS4"
-}
-
-firebase.initializeApp(firebaseConfig);
 
 export async function login(credential: Credential): Promise<string> {
   try {
