+++ server/data-store/src/lib/constants/index.ts
@@ -1,0 +1,12 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyAOVXtaV1OsZp3YWwH_d3fiL2pdbGUuQH8",
+    authDomain: "growbud-50ed4.firebaseapp.com",
+    databaseURL: "https://growbud-50ed4.firebaseio.com",
+    projectId: "growbud-50ed4",
+    storageBucket: "growbud-50ed4.appspot.com",
+    messagingSenderId: "1032847847071",
+    appId: "1:1032847847071:web:116560c04ed237deb9bcaa",
+    measurementId: "G-N5LWZZRVS4"
+}
+
+export default firebaseConfig
